login/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.7
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7